backend/requirements.txt
django
djangorestframework
django-cors-headers
whitenoise
gunicorn
psycopg2-binary
psycopg[binary]
python-dotenv
djangorestframework-simplejwt
